draco_defconfig: Disable exFAT support
The OnePlus implementation of exFAT is dirty, incomplete, and almost
irrelevant for the OnePlus 7 Pro due to the lack of a microSD card slot.
While USB drives could be formatted as exFAT, it is cleaner to just not
include kernel support for the filesystem altogether and wait for an
official CAF merge.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>