draco_defconfig: Enable autogrouping
After talking with some individuals over IRC on the LinuxKernelNoobies
chat, I understand how autogrouping works and why it is significant for
interactive systems.

To dumb it down, assume you have two autogroups that are bound to the
same CPU:

A1:
10 background jobs

A2:
1 foreground job

Without autogrouping, that CPU will give 91% of the time to the
background jobs, with only 9% left for the foreground task.

With autogrouping, both tasks will attempt to get equal time shares, so
that 50% of the CPU time is given to the background jobs and the other
50% are given to the foreground job.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>